+++ dev/tests/layout-shared-crossfade-a-ab.tsx
@@ -9,6 +9,7 @@ const transition = {
 export const App = () => {
     const params = new URLSearchParams(window.location.search)
     const type = params.get("type") || true
+    const size = params.get("size") || false
     const [state, setState] = React.useState(false)
 
     return (
@@ -26,7 +27,7 @@ export const App = () => {
                     id="b"
                     layoutId="box"
                     layout={type}
-                    style={b}
+                    style={size ? aLarge : b}
                     transition={transition}
                     onClick={() => setState(!state)}
                 />
@@ -54,4 +55,12 @@ const b = {
     left: 200,
     width: 300,
     height: 300,
+}
+
+const aLarge = {
+    ...box,
+    top: 100,
+    left: 200,
+    width: 300,
+    height: 600,
 }