+++ packages/framer-motion/cypress/integration/layout-shared.ts
@@ -584,26 +584,78 @@ describe("Shared layout: A -> AB -> A crossfade transition", () => {
                     height: 300,
                 })
             })
-        // .trigger("click")
-        // .wait(50)
-        // .get("#a")
-        // .should(([$box]: any) => {
-        //     expectBbox($box, {
-        //         top: 50,
-        //         left: 100,
-        //         width: 100,
-        //         height: 200,
-        //     })
-        // })
-        // .get("#b")
-        // .should(([$box]: any) => {
-        //     expectBbox($box, {
-        //         top: 50,
-        //         left: 100,
-        //         width: 100,
-        //         height: 200,
-        //     })
-        // })
+    })
+
+    it(`It correctly animates layout="preserve-aspect" as "position" animations if aspect ratios are different`, () => {
+        cy.visit("?test=layout-shared-crossfade-a-ab&type=preserve-aspect")
+            .wait(50)
+            .get("#a")
+            .should(([$box]: any) => {
+                expectBbox($box, {
+                    top: 0,
+                    left: 0,
+                    width: 100,
+                    height: 200,
+                })
+            })
+            .trigger("click")
+            .wait(50)
+            .get("#a")
+            .should(([$box]: any) => {
+                expect(parseFloat($box.style.opacity) || 1).to.equal(1)
+                expectBbox($box, {
+                    top: 50,
+                    left: 100,
+                    width: 100,
+                    height: 200,
+                })
+            })
+            .get("#b")
+            .should(([$box]: any) => {
+                expect(parseFloat($box.style.opacity) || 1).to.equal(1)
+                expectBbox($box, {
+                    top: 50,
+                    left: 100,
+                    width: 300,
+                    height: 300,
+                })
+            })
+    })
+
+    it(`It correctly animates layout="preserve-aspect" as normal layout animations if both aspect ratios are the same`, () => {
+        cy.visit(
+            "?test=layout-shared-crossfade-a-ab&type=preserve-aspect&size=same"
+        )
+            .wait(50)
+            .get("#a")
+            .should(([$box]: any) => {
+                expectBbox($box, {
+                    top: 0,
+                    left: 0,
+                    width: 100,
+                    height: 200,
+                })
+            })
+            .trigger("click")
+            .wait(50)
+            .get("#a")
+            .should(([$box]: any) => {
+                expectBbox($box, {
+                    top: 50,
+                    left: 100,
+                    width: 200,
+                    height: 400,
+                })
+            })
+            .get("#b")
+            .should(([$box]: any) => {
+                expectBbox($box, {
+                    top: 50,
+                    left: 100,
+                    width: 200,
+                    height: 400,
+                })
+            })
     })
 
     it("Correctly fires layout={true} animations after an instant transition", () => {